+++ packages/app-admin-ui/client/pages/Home/index.js
@@ -1,85 +1,55 @@
-import React, { Component, Fragment } from 'react';
-import { withRouter, Link } from 'react-router-dom';
+import React, { Fragment } from 'react';
+import { Link } from 'react-router-dom';
 import { useQuery } from '@apollo/react-hooks';
 
 import { Container, Grid, Cell } from '@arch-ui/layout';
 import { PageTitle } from '@arch-ui/typography';
 
-import CreateItemModal from '../../components/CreateItemModal';
+import { ListProvider } from '../../providers/List';
 import DocTitle from '../../components/DocTitle';
 import PageError from '../../components/PageError';
 import { Box, HeaderInset } from './components';
 import ContainerQuery from '../../components/ContainerQuery';
 import { gqlCountQueries } from '../../classes/List';
 
-class HomePage extends Component {
-  state = { createFromList: null };
+import { useAdminMeta } from '../../providers/AdminMeta';
 
-  openCreateModal = createFromList => event => {
-    event.preventDefault();
-    this.setState({ createFromList });
-  };
-  closeCreateModal = () => this.setState({ createFromList: null });
+const HomePage = ({ lists, data, adminPath }) => (
+  <main>
+    <Container>
+      <HeaderInset>
+        <PageTitle>Dashboard</PageTitle>
+      </HeaderInset>
+      <ContainerQuery>
+        {({ width }) => {
+          let cellWidth = 3;
+          if (width < 1024) cellWidth = 4;
+          if (width < 768) cellWidth = 6;
+          if (width < 480) cellWidth = 12;
+          return (
+            <Grid gap={16}>
+              {lists.map(list => {
+                const { key, path } = list;
+                const meta = data && data[list.gqlNames.listQueryMetaName];
+                return (
+                  <ListProvider list={list} key={key}>
+                    <Cell width={cellWidth}>
+                      <Box to={`${adminPath}/${path}`} meta={meta} />
+                    </Cell>
+                  </ListProvider>
+                );
+              })}
+            </Grid>
+          );
+        }}
+      </ContainerQuery>
+    </Container>
+  </main>
+);
 
-  onCreate = list => ({ data }) => {
-    let { adminPath, history } = this.props;
-    let id = data[list.gqlNames.createMutationName].id;
-    history.push(`${adminPath}/${list.path}/${id}`);
-  };
+const HomepageListProvider = () => {
+  const { getListByKey, listKeys, adminPath } = useAdminMeta();
 
-  render() {
-    const { lists, data, adminPath } = this.props;
-    const { createFromList } = this.state;
-
-    return (
-      <main>
-        <Container>
-          <HeaderInset>
-            <PageTitle>Dashboard</PageTitle>
-          </HeaderInset>
-          <ContainerQuery>
-            {({ width }) => {
-              let cellWidth = 3;
-              if (width < 1024) cellWidth = 4;
-              if (width < 768) cellWidth = 6;
-              if (width < 480) cellWidth = 12;
-
-              return (
-                <Grid gap={16}>
-                  {lists.map(list => {
-                    const { key, path } = list;
-                    const meta = data && data[list.gqlNames.listQueryMetaName];
-
-                    return (
-                      <Fragment key={key}>
-                        <Cell width={cellWidth}>
-                          <Box
-                            list={list}
-                            to={`${adminPath}/${path}`}
-                            meta={meta}
-                            onCreateClick={this.openCreateModal(key)}
-                          />
-                        </Cell>
-                        <CreateItemModal
-                          isOpen={createFromList === key}
-                          list={list}
-                          onClose={this.closeCreateModal}
-                          onCreate={this.onCreate(list)}
-                        />
-                      </Fragment>
-                    );
-                  })}
-                </Grid>
-              );
-            }}
-          </ContainerQuery>
-        </Container>
-      </main>
-    );
-  }
-}
-
-const ListProvider = ({ getListByKey, listKeys, ...props }) => {
   // TODO: A permission query to limit which lists are visible
   const lists = listKeys.map(key => getListByKey(key));
 
@@ -89,7 +59,7 @@ const ListProvider = ({ getListByKey, listKeys, ...props }) => {
   if (lists.length === 0) {
     return (
       <Fragment>
-        <DocTitle>Home</DocTitle>
+        <DocTitle title="Home" />
         <PageError>
           <p>
             No lists defined.{' '}
@@ -134,16 +104,16 @@ const ListProvider = ({ getListByKey, listKeys, ...props }) => {
 
   return (
     <Fragment>
-      <DocTitle>Home</DocTitle>
+      <DocTitle title="Home" />
       {
         // NOTE: `loading` is intentionally omitted here
         // the display of a loading indicator for counts is deferred to the
         // list component so we don't block rendering the lists immediately
         // to the user.
       }
-      <HomePage lists={allowedLists} data={data} {...props} />
+      <HomePage lists={allowedLists} data={data} adminPath={adminPath} />
     </Fragment>
   );
 };
 
-export default withRouter(ListProvider);
+export default HomepageListProvider;