+++ packages/app-graphql/lib/apolloServer.js
@@ -144,8 +144,8 @@ function createApolloServer(keystone, apolloConfig, schemaName, dev) {
   const server = new ApolloServer({
     maxFileSize: 200 * 1024 * 1024,
     maxFiles: 5,
-    ...apolloConfig,
-    ...keystone.getAdminSchema({ schemaName }),
+    typeDefs: keystone.getTypeDefs({ schemaName }),
+    resolvers: keystone.getResolvers({ schemaName }),
     context: ({ req }) => ({
       ...keystone.getGraphQlContext({ schemaName, req }),
       req,
@@ -163,6 +163,7 @@ function createApolloServer(keystone, apolloConfig, schemaName, dev) {
           tracing: dev,
         }),
     formatError: _formatError,
+    ...apolloConfig,
   });
   keystone.registerSchema(schemaName, server.schema);