+++ packages/fields/src/types/Virtual/views/Field.js
@@ -1,12 +1,11 @@
 /** @jsx jsx */
 
 import { jsx } from '@emotion/core';
-import { Component } from 'react';
 
 import { colors, gridSize } from '@arch-ui/theme';
 import { ShieldIcon } from '@arch-ui/icons';
 
-import { FieldContainer } from '@arch-ui/fields';
+import { FieldContainer, FieldDescription } from '@arch-ui/fields';
 import PrettyData from '../prettyData';
 
 export const FieldLabel = props => {
@@ -34,23 +33,19 @@ export const FieldLabel = props => {
   );
 };
 
-export default class VirtualField extends Component {
-  onChange = event => {
-    this.props.onChange(event.target.value);
-  };
+const VirtualField = ({ field, errors, value: serverValue }) => {
+  const value = typeof serverValue !== 'undefined' ? serverValue : '';
+  const canRead = errors.every(
+    error => !(error instanceof Error && error.name === 'AccessDeniedError')
+  );
 
-  render() {
-    const { field, errors, value: serverValue } = this.props;
-    const value = serverValue || '';
-    const canRead = errors.every(
-      error => !(error instanceof Error && error.name === 'AccessDeniedError')
-    );
+  return (
+    <FieldContainer>
+      <FieldLabel field={field} errors={errors} />
+      {field.config.adminDoc && <FieldDescription>{field.config.adminDoc}</FieldDescription>}
+      <PrettyData data={canRead ? value : undefined} />
+    </FieldContainer>
+  );
+};
 
-    return (
-      <FieldContainer>
-        <FieldLabel field={field} errors={errors} />
-        <PrettyData data={canRead ? value : undefined} />
-      </FieldContainer>
-    );
-  }
-}
+export default VirtualField;